AI Powered Tools server/AI_Agent_Framework/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Pillow==10.1.0
requests==2.31.0
